+++ Game/src/audiolib/nodpmi.c
@@ -147,18 +147,18 @@ int DPMI_UnlockMemoryRegion
    return( status );
    }
 
-int DPMI_GetDOSMemory( void **ptr, int *descriptor, unsigned length )
+int DPMI_GetDOSMemory( void **ptr, long *descriptor, unsigned length )
 {
 	/* Lovely... */
 	
 	*ptr = (void *)malloc(length);
 	
-	*descriptor = (int) *ptr;
+	*descriptor = (long) *ptr;
 	
 	return (descriptor == 0) ? DPMI_Error : DPMI_Ok;
 }
 
-int DPMI_FreeDOSMemory( int descriptor )
+int DPMI_FreeDOSMemory( long descriptor )
 {
 	free((void *)descriptor);
